# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Progrid API
  license:
    name: Proprietary
    url: https://progrid.co/legal/terms
  version: "1"
  description: |
    The Progrid public API. The same API powers the console, CLI, Terraform provider,
    SDKs and AI agents (via MCP).

    **Conventions**
    - Base URL: `https://api.progrid.co/v1` (primary). `https://api.progrid.sa/v1` is the same API
      for customers in Saudi Arabia; accounts and tokens work on both.
    - Billing: Progrid Arabia (Saudi Arabia) sells and invoices everything, on both domains, with a
      ZATCA e-invoice in the PRGD-SA series, collected through Moyasar. The team's billing country
      decides currency and VAT: Saudi Arabia pays SAR with 15% VAT; every other country pays USD at
      0% VAT (zero-rated export of services), and those invoices also show the VAT and totals in SAR
      with the exchange rate used. `billingEntity` is always `progrid_arabia` on new data; old
      invoices of Progrid Technologies LLC (`progrid_llc`, PRGD-US) stay readable. See
      `GET /billing/balance` and `GET /geo`.
    - Auth: `Authorization: Bearer prgd_…` (API token) or a console session JWT
    - Lists are cursor-paginated: `?limit=50&cursor=<id>` → `meta.next_cursor`
    - Mutations accept `Idempotency-Key` (replayed for 24h; same key + different body → 409)
    - Long operations return `202 Accepted` and a resource in a transitional status; poll or use webhooks
...
servers:
  - url: https://api.progrid.co/v1
...
security:
  - bearerAuth: []
...
paths:
...
  /pricing:
    get:
      tags: [catalog]
      security: []
      summary: Public price list (SAR book; USD converted at the current exchange rate, 3.75 pegged)
      parameters: [{ name: currency, in: query, schema: { type: string, enum: [USD, SAR], default: USD } }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  currency: { type: string }
                  baseCurrency: { type: string, enum: [SAR] }
                  fxRate: { type: number, description: Book (SAR) to currency rate used for this response }
                  taxRate: { type: number, description: "VAT for customers billed in this currency: 0.15 for SAR (billing country Saudi Arabia), 0 for USD (zero-rated export of services)" }
                  vatRate: { type: number, description: Saudi VAT rate (0.15) }
                  vatNote: { type: string }
                  hoursPerMonth: { type: integer }
                  data: { type: array, items: { type: object, properties: { resourceType: { type: string }, sku: { type: string }, unit: { type: string }, monthlyMinor: { type: integer }, hourlyMinor: { type: integer }, size: { $ref: "#/components/schemas/Size" } } } }
...
  /billing/usage:
    get:
      tags: [billing]
      summary: Rated usage per resource
      parameters: [{ $ref: "#/components/parameters/project" }, { name: from, in: query, schema: { type: string, format: date-time } }, { name: to, in: query, schema: { type: string, format: date-time } }]
      responses: { "200": { description: OK } }
...
  /billing/invoices/{id}:
    get: { tags: [billing], summary: Get an invoice with its usage lines, payments and credit notes, parameters: [{ $ref: "#/components/parameters/id" }], responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Invoice" } } } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    agentKey: { type: http, scheme: bearer, description: "A Connect agent key (prgd_ca_...) for one agent's run API" }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: string } }
    project: { name: project, in: query, description: Project id or slug (default `default`), schema: { type: string } }
...
  schemas:
...
    BillingEntityId:
      type: string
      enum: [progrid_arabia, progrid_llc]
      description: "The contracting and invoicing company: always progrid_arabia on new data. progrid_llc (Progrid Technologies LLC) only appears on old invoices and credit notes (PRGD-US, CN-US). Kept for compatibility."
...
    Invoice:
      type: object
      properties:
        id: { type: string }
        number: { type: string, example: PRGD-SA-2026-00042, description: "PRGD-SA-YYYY-NNNNN; history: PRGD-US-YYYY-NNNNN (Progrid Technologies LLC), PRGD-YYYY-NNNNNN (before 2026-10)" }
        billingEntity: { $ref: "#/components/schemas/BillingEntityId" }
        currency: { type: string, enum: [USD, SAR] }
        periodStart: { type: string, format: date-time }
        periodEnd: { type: string, format: date-time }
        subtotalMinor: { type: integer }
        discountMinor: { type: integer }
        taxMinor: { type: integer, description: "VAT: 15% for billing country SA, 0 otherwise" }
        creditMinor: { type: integer }
        totalMinor: { type: integer }
        creditedMinor: { type: integer }
        status: { type: string, enum: [draft, open, paid, void, uncollectible, credited] }
        eInvoiceType: { type: [string, "null"], example: zatca }
        vatCategory: { type: [string, "null"], enum: [standard, zero_rated_export, null] }
        taxNote: { type: [string, "null"], example: Zero-rated export of services }
        fxRateSar: { type: [string, "null"], example: "3.750000", description: "SAR per one unit of the invoice currency at issue (1 for SAR). Null on invoices issued before 2026-10-12." }
        subtotalSarMinor: { type: [integer, "null"] }
        taxSarMinor: { type: [integer, "null"], description: The VAT amount in SAR (ZATCA) }
        totalSarMinor: { type: [integer, "null"] }
        dueAt: { type: [string, "null"], format: date-time }
        paidAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        creditNotes: { type: array, items: { type: object, properties: { id: { type: string }, number: { type: string, example: CN-SA-2026-00003 }, amountMinor: { type: integer }, reason: { type: string }, createdAt: { type: string, format: date-time } } } }
...
    Size:
      type: object
      properties:
        id: { type: string, example: s-2vcpu-4gb }
        vcpu: { type: integer }
        memoryMb: { type: integer }
        diskGb: { type: integer }
        transferTb: { type: number }
        family: { type: string, enum: [shared, dedicated, gpu] }
...